requirements.txt
# requirements.txt — Pin versions for reproducibility
torch==2.1.0
torchvision==0.16.0
flwr==1.5.0
opacus==1.4.0
librosa==0.10.1
streamlit==1.28.0
tqdm==4.66.1
requests==2.31.0         # for ntfy.sh
numpy==1.24.4
pandas==2.1.1
scikit-learn==1.3.2
matplotlib==3.8.0
pillow==10.1.0
h5py==3.10.0
sentence-transformers==2.2.2   # Semantic Interface (Phase 1)
